requirements.txt
# API Backend
fastapi<1
uvicorn<1

# Core ML
scikit-learn>=1.5.0
numpy>=2.0.0
pandas>=2.2.0

# Explainability
shap>=0.45.0

# Experiment Tracking
mlflow>=2.13.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.22.0

# Web App
streamlit>=1.35.0

# Utilities
joblib>=1.4.0
scipy>=1.13.0

# Testing
pytest>=8.0.0

# Config
pyyaml>=6.0.0
